backend/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.9.0
torch==2.1.0
numpy==1.24.3
python-dotenv==1.0.0
eventlet==0.33.3 